Added flipbook.py to allow for the visualization of changes in PDB structures time taking advantage of the b-factor column.
Added function within the rmsx.py called run_rmsx_flipbook() which runs with all all_chain_rmsx().

I also updated the all_chain_rmsx() function to allow make it return the directory for the combined pdb files or just return the directory with the correct pdbs in it if there was only one chain in the original structure.