Updated so it uses the first frame of the simulation rather than the original topology file. means that RMSD starts at 0.
Also updated the RMSX plot to prevent the offset in the plot caused by it placimg the geo_raster based upon the center of the slice rather than the left or right side of it. previously there was a gap between zero and the middle distance of the first slice.

also added in-line documentation to the plot_rmsx.R